Chatbot/backend/requirements.txt
python-dotenv
fastapi
uvicorn
openai
numpy
scikit-learn
requests